Restructure reports storage
Report files are now stored according to the state of the simulation.
Files are stored in the `running` directory while the simulation is
executed. They are then moved to the `complete` directory when the
simulation finishes successfully. If the simulation fails the files are
moved to the `failed` directory inside a sub-directory with the
timestamp corresponding to the time at which the simulation failed.